add peer to RT & fix/document races
